requirements.txt
itsdangerous==1.1.0
MarkupSafe==1.1.1
Werkzeug==0.15.5
flake8==3.7.9
Flask==1.1.1
flask-cors==3.0.8
gunicorn==20.0.4
h5py==2.10.0
html5lib==1.0.1
httplib2==0.18.1
Jinja2==2.11.1
json5==0.9.1
jsonschema==3.2.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
numpy==1.18.1
oauth2client==4.1.3
oauthlib==3.1.0
pandas==1.0.1
pickle-mixin==1.0.2
pickleshare==0.7.5
requests==2.22.0
requests-oauthlib==1.3.0
tensorflow-cpu==2.2.0
scipy
PIL